[Dispatcher] Update arguments to require argument name (#152)
# Description

Update the dispatcher to require dispatch arguments to have the argument
name as the key of the array.
This also involves updating the routing layer of Http to pass arguments
in from regex matching as string-key/value pairs. So regexes have been
updated
Event dispatching has also been updated to require the event parameter
in dispatches.

## Types of changes

- [ ] Bug fix _(non-breaking change which fixes an issue)_
    <!-- Target the lowest major affected branch -->
- [X] New feature _(non-breaking change which adds functionality)_
    <!-- Target master -->
- [ ] Deprecation _(breaking change which removes functionality)_
    <!-- Target master -->
- [ ] Breaking change _(fix or feature that would cause existing
functionality
  to change)_
<!-- Target master, unless this is a bug fix in which case let's chat
-->
- [ ] Documentation improvement
    <!-- Target appropriate branch -->