[CI] Add cherry-pick workflow (#160)
# Description

This will take over for the merge workflow and allow us to deprecate it.
The merge workflow has issues with merging changes up especially when
changes are made that should stay in those lower version branches.

This will be a safer way to grab changes from lower branches (bugfixes)
and apply them up to higher branches as necessary.

## Types of changes

- [ ] Bug fix _(non-breaking change which fixes an issue)_
    <!-- Target the lowest major affected branch -->
- [X] New feature _(non-breaking change which adds functionality)_
    <!-- Target master -->
- [ ] Deprecation _(breaking change which removes functionality)_
    <!-- Target master -->
- [ ] Breaking change _(fix or feature that would cause existing
functionality
  to change)_
<!-- Target master, unless this is a bug fix in which case let's chat
-->
- [ ] Documentation improvement
    <!-- Target appropriate branch -->