[Http] Update data route object to require dispatch (#180)
# Description

Update data route object to require dispatch, but allow attribute to not
require.

The attribute data object cannot require it because if it did it would
not throw an exception when creating the attribute instance due to
missing required parameter. We auto fill the dispatch based on the
method the route attribute is attached to.

## Types of changes

- [ ] Bug fix _(non-breaking change which fixes an issue)_
    <!-- Target the lowest major affected branch -->
- [ ] New feature _(non-breaking change which adds functionality)_
    <!-- Target master -->
- [X] Deprecation _(breaking change which removes functionality)_
    <!-- Target master -->
- [X] Breaking change _(fix or feature that would cause existing
functionality
  to change)_
<!-- Target master, unless this is a bug fix in which case let's chat
-->
- [ ] Documentation improvement
    <!-- Target appropriate branch -->